Factor AnonymousStorageServer fixture out for re-use
