+++ src/_zkapauthorizer/tests/test_base64.py
@@ -16,6 +16,10 @@
 Tests for ``_zkapauthorizer._base64``.
 """
 
+from __future__ import (
+    absolute_import,
+)
+
 from base64 import (
     urlsafe_b64encode,
 )
