+++ src/_zkapauthorizer/tests/test_controller.py
@@ -16,6 +16,10 @@
 Tests for ``_zkapauthorizer.controller``.
 """
 
+from __future__ import (
+    absolute_import,
+)
+
 from json import (
     loads,
     dumps,
