+++ src/_zkapauthorizer/tests/test_plugin.py
@@ -16,6 +16,10 @@
 Tests for the Tahoe-LAFS plugin.
 """
 
+from __future__ import (
+    absolute_import,
+)
+
 from io import (
     BytesIO,
 )
